Updated the Lose Screens
I put in the new Backgrounds for the Lose Screens in the game. The Restart from Boss scene, Lose Screen and Lose Screen 3 scenes. Don't know if they are being used so I just put in the new Background and buttons in there just in case. The new background is in the folder labelled Win/Lose new in the Background Assets folder and the new buttons are in the Buttons folder